+++ contracts/light-clients/msgs/IUpdateClientMsgs.sol
@@ -19,14 +19,14 @@ interface IUpdateClientMsgs {
     /// @param clientState The client state that was used to verify the header.
     /// @param trustedConsensusState The trusted consensus state.
     /// @param newConsensusState The new consensus state with the verified header.
-    /// @param time The time which the header was verified in seconds.
+    /// @param time The time which the header was verified in nanoseconds.
     /// @param trustedHeight The trusted height.
     /// @param newHeight The new height.
     struct UpdateClientOutput {
         IICS07TendermintMsgs.ClientState clientState;
         IICS07TendermintMsgs.ConsensusState trustedConsensusState;
         IICS07TendermintMsgs.ConsensusState newConsensusState;
-        uint64 time;
+        uint128 time;
         IICS02ClientMsgs.Height trustedHeight;
         IICS02ClientMsgs.Height newHeight;
     }
